Got rid of list of items class
List of items did essentially the same thing as PageScaffold, except it had unneccessary styling around the appbar
